moved convert commands to a separate submodule
